Added the note to make sure "Royal Blue" is used
When applying the Enhanced Dark Theme CSS code, I noticed it broke the UI of Kagi. After some troubleshooting, I found out it was because I had Moon Dark set as the Default Dark Theme, instead of Royal Blue. Changing this allowed the Custom CSS to work properly. Thus, proposing this change to include that information.