Update summarize-page.md
Added the description as to what model summarize page is based on. See https://kagifeedback.org/d/1556-a-transparency-page-or-report-for-kagis-llms/10

Also added the description that the API is a fourth way to access this along with a comparison between the two API models (agnes and muriel). And added some more explanation to the summarize page website part of the document